+++ test-schemas-domain-classes/src/main/scala/testdomains/codepropertygraphminified/accessors/Accessors.scala
@@ -8,13 +8,26 @@ object languagebootstrap extends ConcreteStoredConversions
 object Accessors {
   /* accessors for concrete stored nodes start */
   final class AccessPropertyDispatchType(val node: nodes.StoredNode) extends AnyVal {
-    def dispatchType: String = flatgraph.Accessors.getNodePropertySingle(node.graph, node.nodeKind, 0, node.seq(), "<empty>": String)
+    def dispatchType: String = flatgraph.Accessors.getNodePropertySingle(
+      node.graph,
+      nodeKind = node.nodeKind,
+      propertyKind = 0,
+      seq = node.seq(),
+      default = "<empty>": String
+    )
   }
   final class AccessPropertyName(val node: nodes.StoredNode) extends AnyVal {
-    def name: String = flatgraph.Accessors.getNodePropertySingle(node.graph, node.nodeKind, 1, node.seq(), "<empty>": String)
+    def name: String = flatgraph.Accessors.getNodePropertySingle(
+      node.graph,
+      nodeKind = node.nodeKind,
+      propertyKind = 1,
+      seq = node.seq(),
+      default = "<empty>": String
+    )
   }
   final class AccessPropertyOrder(val node: nodes.StoredNode) extends AnyVal {
-    def order: Int = flatgraph.Accessors.getNodePropertySingle(node.graph, node.nodeKind, 2, node.seq(), -1: Int)
+    def order: Int =
+      flatgraph.Accessors.getNodePropertySingle(node.graph, nodeKind = node.nodeKind, propertyKind = 2, seq = node.seq(), default = -1: Int)
   }
   /* accessors for concrete stored nodes end */
