+++ test-schemas-domain-classes/src/main/scala/testdomains/codepropertygraphminified/nodes/Method.scala
@@ -49,73 +49,3 @@ class Method(graph_4762: flatgraph.Graph, seq_4762: Int)
 
   override def canEqual(that: Any): Boolean = that != null && that.isInstanceOf[Method]
 }
-
-object NewMethod {
-  def apply(): NewMethod                             = new NewMethod
-  private val outNeighbors: Map[String, Set[String]] = Map()
-  private val inNeighbors: Map[String, Set[String]]  = Map("CALL" -> Set("CALL"))
-
-  object InsertionHelpers {
-    object NewNodeInserter_Method_name extends flatgraph.NewNodePropertyInsertionHelper {
-      override def insertNewNodeProperties(newNodes: mutable.ArrayBuffer[flatgraph.DNode], dst: AnyRef, offsets: Array[Int]): Unit = {
-        if (newNodes.isEmpty) return
-        val dstCast = dst.asInstanceOf[Array[String]]
-        val seq     = newNodes.head.storedRef.get.seq()
-        var offset  = offsets(seq)
-        var idx     = 0
-        while (idx < newNodes.length) {
-          val nn = newNodes(idx)
-          nn match {
-            case generated: NewMethod =>
-              dstCast(offset) = generated.name
-              offset += 1
-            case _ =>
-          }
-          assert(seq + idx == nn.storedRef.get.seq(), "internal consistency check")
-          idx += 1
-          offsets(idx + seq) = offset
-        }
-      }
-    }
-  }
-}
-
-class NewMethod extends NewNode(1) with MethodBase with DeclarationNew {
-  override type StoredNodeType = Method
-  override def label: String = "METHOD"
-
-  override def isValidOutNeighbor(edgeLabel: String, n: NewNode): Boolean = {
-    NewMethod.outNeighbors.getOrElse(edgeLabel, Set.empty).contains(n.label)
-  }
-  override def isValidInNeighbor(edgeLabel: String, n: NewNode): Boolean = {
-    NewMethod.inNeighbors.getOrElse(edgeLabel, Set.empty).contains(n.label)
-  }
-
-  var name: String                   = "<empty>": String
-  def name(value: String): this.type = { this.name = value; this }
-  override def countAndVisitProperties(interface: flatgraph.BatchedUpdateInterface): Unit = {
-    interface.countProperty(this, 1, 1)
-  }
-
-  override def copy: this.type = {
-    val newInstance = new NewMethod
-    newInstance.name = this.name
-    newInstance.asInstanceOf[this.type]
-  }
-
-  override def productElementName(n: Int): String =
-    n match {
-      case 0 => "name"
-      case _ => ""
-    }
-
-  override def productElement(n: Int): Any =
-    n match {
-      case 0 => this.name
-      case _ => null
-    }
-
-  override def productPrefix                = "NewMethod"
-  override def productArity                 = 1
-  override def canEqual(that: Any): Boolean = that != null && that.isInstanceOf[NewMethod]
-}
