+++ test-schemas-domain-classes/src/main/scala/testdomains/codepropertygraphminified/nodes/RootTypes.scala
@@ -17,10 +17,16 @@ abstract class StoredNode(graph_4762: flatgraph.Graph, kind_4762: Short, seq_476
     extends flatgraph.GNode(graph_4762, kind_4762, seq_4762)
     with AbstractNode {
 
-  final def _callOut: Iterator[StoredNode] =
-    flatgraph.Accessors.getNeighborsOut(this.graph, this.nodeKind, this.seq, 0).asInstanceOf[Iterator[StoredNode]]
-  final def _callIn: Iterator[StoredNode] =
-    flatgraph.Accessors.getNeighborsIn(this.graph, this.nodeKind, this.seq, 0).asInstanceOf[Iterator[StoredNode]]
+  final def _callOut: Iterator[StoredNode] = {
+    flatgraph.Accessors
+      .getNeighborsOut(this.graph, nodeKind = this.nodeKind, seq = this.seq, edgeKind = 0)
+      .asInstanceOf[Iterator[StoredNode]]
+  }
+  final def _callIn: Iterator[StoredNode] = {
+    flatgraph.Accessors
+      .getNeighborsIn(this.graph, nodeKind = this.nodeKind, seq = this.seq, edgeKind = 0)
+      .asInstanceOf[Iterator[StoredNode]]
+  }
 
 }
